refactor: move login button to utils
